refactor: remove oracle from h1st, move sgboost to model.ml
